requirements.txt
requests
dotenv
openai
pyyaml
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
